// frontend/src/api/stub/models/Order.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { OrderStatus } from './OrderStatus';
import {
    OrderStatusFromJSON,
    OrderStatusFromJSONTyped,
    OrderStatusToJSON,
    OrderStatusToJSONTyped,
} from './OrderStatus';

/**
 * Order database table
 * @export
 * @interface Order
 */
export interface Order {
    /**
     * 
     * @type {number}
     * @memberof Order
     */
    id?: number | null;
    /**
     * 
     * @type {number}
     * @memberof Order
     */
    tableId: number;
    /**
     * 
     * @type {number}
     * @memberof Order
     */
    userId: number;
    /**
     * 
     * @type {OrderStatus}
     * @memberof Order
     */
    status?: OrderStatus;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Order
     */
    closedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof Order
     */
    notes?: string | null;
}



/**
 * Check if a given object implements the Order interface.
 */
export function instanceOfOrder(value: object): value is Order {
    if (!('tableId' in value) || value['tableId'] === undefined) return false;
    if (!('userId' in value) || value['userId'] === undefined) return false;
    return true;
}

export function OrderFromJSON(json: any): Order {
    return OrderFromJSONTyped(json, false);
}

export function OrderFromJSONTyped(json: any, ignoreDiscriminator: boolean): Order {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'tableId': json['table_id'],
        'userId': json['user_id'],
        'status': json['status'] == null ? undefined : OrderStatusFromJSON(json['status']),
        'createdAt': json['created_at'] == null ? undefined : (new Date(json['created_at'])),
        'closedAt': json['closed_at'] == null ? undefined : (new Date(json['closed_at'])),
        'notes': json['notes'] == null ? undefined : json['notes'],
    };
}

export function OrderToJSON(json: any): Order {
    return OrderToJSONTyped(json, false);
}

export function OrderToJSONTyped(value?: Order | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'table_id': value['tableId'],
        'user_id': value['userId'],
        'status': OrderStatusToJSON(value['status']),
        'created_at': value['createdAt'] == null ? value['createdAt'] : value['createdAt'].toISOString(),
        'closed_at': value['closedAt'] == null ? value['closedAt'] : value['closedAt'].toISOString(),
        'notes': value['notes'],
    };
}