// frontend/src/api/stub/models/OrderUpdate.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { OrderStatus } from './OrderStatus';
import {
    OrderStatusFromJSON,
    OrderStatusFromJSONTyped,
    OrderStatusToJSON,
    OrderStatusToJSONTyped,
} from './OrderStatus';

/**
 * Schema for updating an order (all fields optional for PATCH)
 * @export
 * @interface OrderUpdate
 */
export interface OrderUpdate {
    /**
     * 
     * @type {OrderStatus}
     * @memberof OrderUpdate
     */
    status?: OrderStatus | null;
    /**
     * 
     * @type {Date}
     * @memberof OrderUpdate
     */
    closedAt?: Date | null;
}



/**
 * Check if a given object implements the OrderUpdate interface.
 */
export function instanceOfOrderUpdate(value: object): value is OrderUpdate {
    return true;
}

export function OrderUpdateFromJSON(json: any): OrderUpdate {
    return OrderUpdateFromJSONTyped(json, false);
}

export function OrderUpdateFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderUpdate {
    if (json == null) {
        return json;
    }
    return {
        
        'status': json['status'] == null ? undefined : OrderStatusFromJSON(json['status']),
        'closedAt': json['closed_at'] == null ? undefined : (new Date(json['closed_at'])),
    };
}

export function OrderUpdateToJSON(json: any): OrderUpdate {
    return OrderUpdateToJSONTyped(json, false);
}

export function OrderUpdateToJSONTyped(value?: OrderUpdate | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'status': OrderStatusToJSON(value['status']),
        'closed_at': value['closedAt'] == null ? value['closedAt'] : value['closedAt'].toISOString(),
    };
}