// frontend/src/api/stub/models/OrderStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Restaurant POS
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const OrderStatus = {
    Draft: 'draft',
    InProgress: 'in_progress',
    AwaitingPayment: 'awaiting_payment',
    OnHold: 'on_hold',
    Completed: 'completed',
    Cancelled: 'cancelled',
    Voided: 'voided',
    Refunded: 'refunded'
} as const;
export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export function instanceOfOrderStatus(value: any): boolean {
    for (const key in OrderStatus) {
        if (Object.prototype.hasOwnProperty.call(OrderStatus, key)) {
            if (OrderStatus[key as keyof typeof OrderStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function OrderStatusFromJSON(json: any): OrderStatus {
    return OrderStatusFromJSONTyped(json, false);
}

export function OrderStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderStatus {
    return json as OrderStatus;
}

export function OrderStatusToJSON(value?: OrderStatus | null): any {
    return value as any;
}

export function OrderStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): OrderStatus {
    return value as OrderStatus;
}